why debug not works

+++ lib/Thermocouple/Cryo_Thermocouple.h
@@ -12,14 +12,14 @@
 namespace Thermocouple {
   class Cryo {
     public:
-      int init(int numSensors, Adafruit_MCP9600 *cryo_boards, int * addrs, _themotype * types, TwoWire *theWire); // assume that numSensors is < max Size of packet. Add some error checking here
+      int init(uint8_t numSensors, Adafruit_MCP9600 *cryo_boards, uint8_t * addrs, _themotype * types, TwoWire *theWire); // assume that numSensors is < max Size of packet. Add some error checking here
       void readCryoTemps(float *data);
-      void readSpecificCryoTemp(int index, float *data);
+      void readSpecificCryoTemp(uint8_t index, float *data);
       int freeAllResources();
     private:
       Adafruit_MCP9600 * _cryo_amp_boards;
-      int * _addrs;
-      int _numSensors;
+      uint8_t * _addrs;
+      uint8_t _numSensors;
       float * _latestReads;
   };
